+++ src/views/dashboard/layout.pug
@@ -49,4 +49,4 @@ html
 
         block js-imports
 
-            script(src="https://unpkg.com/alpinejs@3.x.x")
+        script(src="https://unpkg.com/alpinejs@3.x.x")
