+++ src/views/dashboard/pages/products.pug
@@ -154,7 +154,7 @@ block layout-content
                             input(type="text" name="brand" required).form-control
 
                         .col
-                            label(for="os").form-label Operating System
+                            label(for="os").form-label OS
                             input(type="text" name="os" required).form-control
                         
                         .col
