+++ src/views/dashboard/pages/users.pug
@@ -38,7 +38,8 @@ block layout-content
               label.form-label(for="userRole") Role
               select.form-select(name="role" id="userRole")
                 option(value="administrator") Administrator
-                option(value="user") User
+                option(value="manager") Manager
+                option(value="customer") Customer
           .modal-footer
             button.btn.btn-primary(type="submit") Save
             button.btn.btn-secondary(type="button" data-bs-dismiss="modal") Cancel
