+++ src/views/dashboard/partials/sidebar.pug
@@ -15,6 +15,7 @@
                 `).nav.nav-pills.nav-fill.w-100
                 li.nav-item.text-start
                     each [href, nav] of Object.entries(navs)
-                        a(href=`/dashboard${href}` class=(activeNav === href ? "active" : "")).nav-link
-                            iconify-icon(icon=`mynaui:${nav.icon}`)
-                            span.ms-3 #{nav.title}
+                        if href !== "/users" || roleName === "administrator"
+                            a(href=`/dashboard${href}` class=(activeNav === href ? "active" : "")).nav-link
+                                iconify-icon(icon=`mynaui:${nav.icon}`)
+                                span.ms-3 #{nav.title}
