+++ src/views/store/pages/cart.pug
@@ -12,8 +12,8 @@ block layout-content
                 p.text-muted.text-center.mb-4 Find the perfect phone to keep it company!
                 
                 // Primary CTA button
-                a.btn.btn-primary.btn-lg.px-4.py-2.shop-now-btn(href="/#bestSellers")
-                    iconify-icon(icon="mynaui:smartphone-add" style="vertical-align: middle; margin-right: 8px;")
+                a.btn.btn-dark.btn-lg.px-4.py-2.shop-now-btn(href="/#bestSellers")
+                    iconify-icon(icon="mynaui:cart" style="margin-right: 8px;")
                     | Shop top phones
         else
             h1.mb-5 Your cart
